Split win and unix into different files
